Add exclude_members, and switch from string to stream
This adds a new request parameter to our Slack channels call to not
return channel members. This information is which Slack users are
currently in that channel. We don't use that information and it just
makes the response larger and more expensive to parse. This commit
also switches from a string response body to an InputStream
response. This allows us to avoid the overhead of the full response
being a string in memory before passing it to our JSON parser.

Fixes #8733